transformer ada hvd

+++ scripts/machine_translation/train_transformer_ada_hvd.py
@@ -348,7 +348,7 @@ def train():
             if rank==0:
                 logging.info('[Epoch {} Batch {}/{}], src_wc={}, tgt_wc={}, bs={}'.format(epoch_id, batch_id + 1, len(train_data_loader), 
                                                                                                  debug_array_1_np[0], debug_array_1_np[1], debug_array_1_np[2]))
-            debug_array_2 = mx.nd.zeros((num_workers, ))
+            debug_array_2 = mx.nd.zeros((num_workers, ), dtype='float32')
             debug_array_2[rank] = float(ls.asscalar())
             hvd.allreduce_(debug_array_2, average=False, name='debug_array_2', priority=0)
             debug_array_2_np = debug_array_2.asnumpy()
